Allows using a pure modular build.
This allows the user to force modules to be used. Since the std module
is enable it's not needed to include the Standard library headers.
Clang 17 does not work properly in this case, so allow users to avoid
these includes.

Note static_asserts can't be exported so they are disabled.